move UserListDetailsPage and LearningPathDetailsPage into the ListDetailsPage folder
