+++ frontends/api/package.json
@@ -6,7 +6,7 @@
     ".": "./src/generated/v1/api.ts",
     "./v0": "./src/generated/v0/api.ts",
     "./hooks/*": "./src/hooks/*/index.ts",
-    "./constants/*": "./src/common/constants.ts",
+    "./constants": "./src/common/constants.ts",
     "./test-utils/factories": "./src/test-utils/factories/index.ts",
     "./test-utils": "./src/test-utils/index.ts"
   },
